[hip-flex & phase] Change phase input as int (coherence with PG)

+++ include/sot/talos_balance/hip-flexibility-compensation.hh
@@ -60,7 +60,7 @@ class HIPFLEXIBILITYCOMPENSATION_EXPORT HipFlexibilityCompensation : public ::dy
 
   /* --- SIGNALS --- */
   /// \brief Walking phase
-  DECLARE_SIGNAL_IN(phase, double);
+  DECLARE_SIGNAL_IN(phase, int);
 
   /// \brief  Desired joint configuration of the robot
   DECLARE_SIGNAL_IN(q_des, dynamicgraph::Vector);
